+++ src/main/java/teammates/ui/webapi/action/DeleteCourseAction.java
@@ -1,16 +1,14 @@
 package teammates.ui.webapi.action;
 
-import org.apache.http.HttpStatus;
-
-import teammates.common.exception.EntityDoesNotExistException;
-import teammates.common.exception.InvalidParametersException;
 import teammates.common.exception.UnauthorizedAccessException;
 import teammates.common.util.Const;
 
 /**
- * Action: Move a course to Recycle Bin (soft-delete) for an instructor.
+ * Action: Permanently deletes a course from Recycle Bin for an instructor.
  */
 public class DeleteCourseAction extends Action {
+
+    private String idOfCourseToDelete;
 
     @Override
     protected AuthType getMinAuthLevel() {
@@ -22,25 +20,20 @@ public class DeleteCourseAction extends Action {
         if (!userInfo.isInstructor) {
             throw new UnauthorizedAccessException("Instructor privilege is required to access this resource.");
         }
-
-        String idOfCourseToDelete = getNonNullRequestParamValue(Const.ParamsNames.COURSE_ID);
+        idOfCourseToDelete = getNonNullRequestParamValue(Const.ParamsNames.COURSE_ID);
         gateKeeper.verifyAccessible(logic.getInstructorForGoogleId(idOfCourseToDelete, userInfo.id),
-                logic.getCourse(idOfCourseToDelete), Const.ParamsNames.INSTRUCTOR_PERMISSION_MODIFY_COURSE);
+                logic.getCourse(idOfCourseToDelete),
+                Const.ParamsNames.INSTRUCTOR_PERMISSION_MODIFY_COURSE);
     }
 
     @Override
     public ActionResult execute() {
-        String idOfCourseToDelete = getNonNullRequestParamValue(Const.ParamsNames.COURSE_ID);
-        try {
-            logic.moveCourseToRecycleBin(idOfCourseToDelete);
 
-            return new JsonResult("The course " + idOfCourseToDelete
-                    + " has been deleted. You can restore it from the Recycle Bin manually.");
-        } catch (InvalidParametersException e) {
-            return new JsonResult(e.getMessage(), HttpStatus.SC_BAD_REQUEST);
-        } catch (EntityDoesNotExistException e) {
-            return new JsonResult(e.getMessage(), HttpStatus.SC_INTERNAL_SERVER_ERROR);
-        }
+        idOfCourseToDelete = getNonNullRequestParamValue(Const.ParamsNames.COURSE_ID);
+
+        logic.deleteCourse(idOfCourseToDelete);
+
+        String statusMessage = "The course " + idOfCourseToDelete + " has been permanently deleted.";
+        return new JsonResult(statusMessage);
     }
-
 }
